# tools/sdk-ts/openapi.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Notifo API",
    "version": "1.0.0",
    "x-logo": {
      "url": "/logo.svg"
    }
  },
  "servers": [
    {
      "url": "https://localhost:5002"
    }
  ],
  "paths": {
...
    "/api/apps/{appId}/media": {
      "get": {
        "tags": [
          "Media"
        ],
        "summary": "Query media items.",
        "operationId": "Media_GetMedias",
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "required": true,
            "description": "The app where the media belongs to.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "query",
            "in": "query",
            "description": "The optional query to search for items.",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 2
          },
          {
            "name": "take",
            "in": "query",
            "description": "The number of items to return.",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 3
          },
          {
            "name": "skip",
            "in": "query",
            "description": "The number of items to skip.",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 4
          }
        ],
        "responses": {
          "200": {
            "description": "Media returned.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponseDtoOfMediaDto"
                }
              }
            }
          },
          "404": {
            "description": "App not found."
          },
          "500": {
            "description": "Operation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "x-method-name": "getMedias"
      },
      "post": {
        "tags": [
          "Media"
        ],
        "summary": "Upload a media object.",
        "operationId": "Media_Upload",
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "required": true,
            "description": "The app id where the media belongs to.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Media uploaded."
          },
          "404": {
            "description": "App not found."
          },
          "400": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Operation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "x-method-name": "upload"
      }
    },
    "/api/apps/{appId}/media/{fileName}": {
      "get": {
        "tags": [
          "Media"
        ],
        "summary": "Download a media object.",
        "operationId": "Media_Download",
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "required": true,
            "description": "The app id where the media belongs to.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "fileName",
            "in": "path",
            "required": true,
            "description": "The name of the media to download.",
            "schema": {
              "type": "string"
            },
            "x-position": 2
          },
          {
            "name": "cache",
            "in": "query",
            "description": "The cache duration.",
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "x-position": 3
          },
          {
            "name": "download",
            "in": "query",
            "description": "Set it to 1 to create a download response.",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 4
          },
          {
            "name": "bg",
            "in": "query",
            "description": "Optional background color.",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 5
          },
          {
            "name": "width",
            "in": "query",
            "description": "The target width when an image.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "x-position": 6
          },
          {
            "name": "height",
            "in": "query",
            "description": "The target height when an image.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "x-position": 7
          },
          {
            "name": "quality",
            "in": "query",
            "description": "The target quality when an image.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "x-position": 8
          },
          {
            "name": "preset",
            "in": "query",
            "description": "A preset dimension.",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 9
          },
          {
            "name": "mode",
            "in": "query",
            "description": "The resize mode.",
            "schema": {
              "nullable": true,
              "$ref": "#/components/schemas/ResizeMode"
            },
            "x-position": 10
          },
          {
            "name": "focusX",
            "in": "query",
            "description": "The x position of the focues point.",
            "schema": {
              "type": "number",
              "format": "float",
              "nullable": true
            },
            "x-position": 11
          },
          {
            "name": "focusY",
            "in": "query",
            "description": "The y position of the focues point.",
            "schema": {
              "type": "number",
              "format": "float",
              "nullable": true
            },
            "x-position": 12
          },
          {
            "name": "force",
            "in": "query",
            "description": "True to resize it and clear the cache.",
            "schema": {
              "type": "boolean"
            },
            "x-position": 13
          },
          {
            "name": "emptyOnFailure",
            "in": "query",
            "description": "True, to return an empty image on failure.",
            "schema": {
              "type": "boolean"
            },
            "x-position": 14
          }
        ],
        "responses": {
          "200": {
            "description": "Media returned.",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Media or app not found."
          },
          "500": {
            "description": "Operation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "x-method-name": "download"
      },
      "delete": {
        "tags": [
          "Media"
        ],
        "summary": "Delete a media.",
        "operationId": "Media_Delete",
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "required": true,
            "description": "The app id where the media belongs to.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "fileName",
            "in": "path",
            "required": true,
            "description": "The file name of the media.",
            "schema": {
              "type": "string"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "204": {
            "description": "Media deleted."
          },
          "404": {
            "description": "App not found."
          },
          "400": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          },
          "500": {
            "description": "Operation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "x-method-name": "delete"
      }
    },
    "/api/asset/{appId}/{fileName}": {
      "get": {
        "tags": [
          "Media"
        ],
        "summary": "Download a media object.",
        "operationId": "Media_Download2",
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "required": true,
            "description": "The app id where the media belongs to.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "fileName",
            "in": "path",
            "required": true,
            "description": "The name of the media to download.",
            "schema": {
              "type": "string"
            },
            "x-position": 2
          },
          {
            "name": "cache",
            "in": "query",
            "description": "The cache duration.",
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "x-position": 3
          },
          {
            "name": "download",
            "in": "query",
            "description": "Set it to 1 to create a download response.",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 4
          },
          {
            "name": "bg",
            "in": "query",
            "description": "Optional background color.",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 5
          },
          {
            "name": "width",
            "in": "query",
            "description": "The target width when an image.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "x-position": 6
          },
          {
            "name": "height",
            "in": "query",
            "description": "The target height when an image.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "x-position": 7
          },
          {
            "name": "quality",
            "in": "query",
            "description": "The target quality when an image.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "x-position": 8
          },
          {
            "name": "preset",
            "in": "query",
            "description": "A preset dimension.",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 9
          },
          {
            "name": "mode",
            "in": "query",
            "description": "The resize mode.",
            "schema": {
              "nullable": true,
              "$ref": "#/components/schemas/ResizeMode"
            },
            "x-position": 10
          },
          {
            "name": "focusX",
            "in": "query",
            "description": "The x position of the focues point.",
            "schema": {
              "type": "number",
              "format": "float",
              "nullable": true
            },
            "x-position": 11
          },
          {
            "name": "focusY",
            "in": "query",
            "description": "The y position of the focues point.",
            "schema": {
              "type": "number",
              "format": "float",
              "nullable": true
            },
            "x-position": 12
          },
          {
            "name": "force",
            "in": "query",
            "description": "True to resize it and clear the cache.",
            "schema": {
              "type": "boolean"
            },
            "x-position": 13
          },
          {
            "name": "emptyOnFailure",
            "in": "query",
            "description": "True, to return an empty image on failure.",
            "schema": {
              "type": "boolean"
            },
            "x-position": 14
          }
        ],
        "responses": {
          "200": {
            "description": "Media returned.",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Media or app not found."
          },
          "500": {
            "description": "Operation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "x-method-name": "download2"
      }
    },
    "/api/assets/{appId}/{fileName}": {
      "get": {
        "tags": [
          "Media"
        ],
        "summary": "Download a media object.",
        "operationId": "Media_Download3",
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "required": true,
            "description": "The app id where the media belongs to.",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          },
          {
            "name": "fileName",
            "in": "path",
            "required": true,
            "description": "The name of the media to download.",
            "schema": {
              "type": "string"
            },
            "x-position": 2
          },
          {
            "name": "cache",
            "in": "query",
            "description": "The cache duration.",
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "x-position": 3
          },
          {
            "name": "download",
            "in": "query",
            "description": "Set it to 1 to create a download response.",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 4
          },
          {
            "name": "bg",
            "in": "query",
            "description": "Optional background color.",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 5
          },
          {
            "name": "width",
            "in": "query",
            "description": "The target width when an image.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "x-position": 6
          },
          {
            "name": "height",
            "in": "query",
            "description": "The target height when an image.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "x-position": 7
          },
          {
            "name": "quality",
            "in": "query",
            "description": "The target quality when an image.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "nullable": true
            },
            "x-position": 8
          },
          {
            "name": "preset",
            "in": "query",
            "description": "A preset dimension.",
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 9
          },
          {
            "name": "mode",
            "in": "query",
            "description": "The resize mode.",
            "schema": {
              "nullable": true,
              "$ref": "#/components/schemas/ResizeMode"
            },
            "x-position": 10
          },
          {
            "name": "focusX",
            "in": "query",
            "description": "The x position of the focues point.",
            "schema": {
              "type": "number",
              "format": "float",
              "nullable": true
            },
            "x-position": 11
          },
          {
            "name": "focusY",
            "in": "query",
            "description": "The y position of the focues point.",
            "schema": {
              "type": "number",
              "format": "float",
              "nullable": true
            },
            "x-position": 12
          },
          {
            "name": "force",
            "in": "query",
            "description": "True to resize it and clear the cache.",
            "schema": {
              "type": "boolean"
            },
            "x-position": 13
          },
          {
            "name": "emptyOnFailure",
            "in": "query",
            "description": "True, to return an empty image on failure.",
            "schema": {
              "type": "boolean"
            },
            "x-position": 14
          }
        ],
        "responses": {
          "200": {
            "description": "Media returned.",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "404": {
            "description": "Media or app not found."
          },
          "500": {
            "description": "Operation failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDto"
                }
              }
            }
          }
        },
        "x-method-name": "download3"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "statusCode"
        ],
        "properties": {
          "message": {
            "type": "string",
            "description": "The error message.",
            "nullable": true
          },
          "errorCode": {
            "type": "string",
            "description": "The error code.",
            "nullable": true
          },
          "traceId": {
            "type": "string",
            "description": "The optional trace id.",
            "nullable": true
          },
          "type": {
            "type": "string",
            "description": "The error type, usually a link.",
            "nullable": true
          },
          "details": {
            "type": "array",
            "description": "Detailed error messages.",
            "nullable": true,
            "items": {
              "type": "string"
            }
          },
          "statusCode": {
            "type": "integer",
            "description": "Status code of the http response.",
            "format": "int32"
          }
        }
      },
...
      "ListResponseDtoOfMediaDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items",
          "total"
        ],
        "properties": {
          "items": {
            "type": "array",
            "description": "The items.",
            "items": {
              "$ref": "#/components/schemas/MediaDto"
            }
          },
          "total": {
            "type": "integer",
            "description": "The total number of items.",
            "format": "int64"
          }
        }
      },
      "MediaDto": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "mimeType",
          "fileName",
          "fileInfo",
          "url",
          "fileSize",
          "created",
          "lastUpdate",
          "type",
          "metadata"
        ],
        "properties": {
          "mimeType": {
            "type": "string",
            "description": "The mime type."
          },
          "fileName": {
            "type": "string",
            "description": "The file name."
          },
          "fileInfo": {
            "type": "string",
            "description": "Generated information about the file."
          },
          "url": {
            "type": "string",
            "description": "The url to the media item."
          },
          "fileSize": {
            "type": "integer",
            "description": "The size of the media file.",
            "format": "int64"
          },
          "created": {
            "type": "string",
            "description": "The date time (ISO 8601) when the media has been created.",
            "format": "date-time"
          },
          "lastUpdate": {
            "type": "string",
            "description": "The date time (ISO 8601) when the media has been updated.",
            "format": "date-time"
          },
          "type": {
            "description": "The type of the media.",
            "$ref": "#/components/schemas/MediaType"
          },
          "metadata": {
            "description": "Metadata about the media.",
            "$ref": "#/components/schemas/MediaMetadata"
          }
        }
      },
      "MediaType": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "Unknown",
          "Image",
          "Audio",
          "Video"
        ],
        "enum": [
          "Unknown",
          "Image",
          "Audio",
          "Video"
        ]
      },
      "MediaMetadata": {
        "type": "object",
        "additionalProperties": {
          "type": "string"
        }
      },
      "ResizeMode": {
        "type": "string",
        "description": "",
        "x-enumNames": [
          "Crop",
          "CropUpsize",
          "Pad",
          "BoxPad",
          "Max",
          "Min",
          "Stretch"
        ],
        "enum": [
          "Crop",
          "CropUpsize",
          "Pad",
          "BoxPad",
          "Max",
          "Min",
          "Stretch"
        ]
      },
...
    }
  }
}